+++ Tests/EFCore.Lazy.Loading.Tests/EFCore.Lazy.Loading.Tests.Infrastructure/Persistence/ApplicationDbContext.cs
@@ -71,7 +71,10 @@ namespace EFCore.Lazy.Loading.Tests.Infrastructure.Persistence
                     .SelectMany(x => x)
                     .FirstOrDefault(domainEvent => !domainEvent.IsPublished);
 
-                if (domainEventEntity == null) break;
+                if (domainEventEntity is null)
+                {
+                    break;
+                }
 
                 domainEventEntity.IsPublished = true;
                 await _domainEventService.Publish(domainEventEntity, cancellationToken);