+++ Tests/MassTransit.RabbitMQ/MassTransit.RabbitMQ.Infrastructure/Persistence/ApplicationDbContext.cs
@@ -66,7 +66,10 @@ namespace MassTransit.RabbitMQ.Infrastructure.Persistence
                     .SelectMany(x => x)
                     .FirstOrDefault(domainEvent => !domainEvent.IsPublished);
 
-                if (domainEventEntity == null) break;
+                if (domainEventEntity is null)
+                {
+                    break;
+                }
 
                 domainEventEntity.IsPublished = true;
                 await _domainEventService.Publish(domainEventEntity, cancellationToken);