+++ Tests/Publish.CleanArch.MassTransit.OutboxEF.TestApplication/Publish.CleanArch.MassTransit.OutboxEF.TestApplication.Infrastructure/Persistence/ApplicationDbContext.cs
@@ -76,7 +76,10 @@ namespace Publish.CleanArch.MassTransit.OutboxEF.TestApplication.Infrastructure.
                     .SelectMany(x => x)
                     .FirstOrDefault(domainEvent => !domainEvent.IsPublished);
 
-                if (domainEventEntity == null) break;
+                if (domainEventEntity is null)
+                {
+                    break;
+                }
 
                 domainEventEntity.IsPublished = true;
                 await _domainEventService.Publish(domainEventEntity, cancellationToken);