+++ Tests/Publish.CleanArch.MassTransit.OutboxNone.TestApplication/Publish.CleanArch.MassTransit.OutboxNone.TestApplication.Infrastructure/Persistence/ApplicationDbContext.cs
@@ -71,7 +71,10 @@ namespace Publish.CleanArch.MassTransit.OutboxNone.TestApplication.Infrastructur
                     .SelectMany(x => x)
                     .FirstOrDefault(domainEvent => !domainEvent.IsPublished);
 
-                if (domainEventEntity == null) break;
+                if (domainEventEntity is null)
+                {
+                    break;
+                }
 
                 domainEventEntity.IsPublished = true;
                 await _domainEventService.Publish(domainEventEntity, cancellationToken);